Try to fix compiler warning. Note we must use long here. Otherwise the protect call to rb_ary_entry a few lines below becomes ambiguous.

+++ rice/stl/vector.ipp
@@ -368,7 +368,7 @@ namespace Rice
     template<typename T>
     std::vector<T> vectorFromArray(VALUE value)
     {
-      size_t length = protect(rb_array_len, value);
+      long length = protect(rb_array_len, value);
       std::vector<T> result(length);
 
       for (long i = 0; i < length; i++)
